pyproject.toml
[project]
name = "dataset-cleaner"
version = "0.1.0"
description = "CLI tool to validate, clean and report messy CSV datasets"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "pandas>=2.0",
    "pydantic>=2.0",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "scipy>=1.12",
    "ollama>=0.3",
]

[project.scripts]
dataset-cleaner = "dataset_cleaner.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataset_cleaner"]